+++ .idea/encodings.xml
@@ -5,15 +5,19 @@
     <file url="file://$PROJECT_DIR$/TESTdemo/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/gzhuoj-common/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/gzhuoj-common/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-config/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-config/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-gateway/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-gateway/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/gzhuoj-microservice-core/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/gzhuoj-microservice-core/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/gzhuoj-contest/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/gzhuoj-contest/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/gzhuoj-problem/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/gzhuoj-problem/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/gzhuoj-user/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/gzhuoj-user/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/gzhuoj-microservices/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/gzhuoj-contest/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/gzhuoj-contest/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/gzhuoj-problem/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/gzhuoj-problem/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/gzhuoj-user/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/gzhuoj-user/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/gzhuoj-services/src/main/resources" charset="UTF-8" />
   </component>
 </project>
