+++ gzhuoj-gateway/src/main/resources/application.yaml
@@ -1,0 +1,5 @@
+spring:
+  application:
+    name: gzhuoj-gateway
+  profiles:
+    active: dev
