--- gzhuoj-microservices/gzhuoj-problem/src/main/java/com/gzhuoj/problem/ProblemApplication.java
@@ -1,13 +1,0 @@
-package com.gzhuoj.problem;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@MapperScan("com.gzhuoj.problem.mapper")
-public class ProblemApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ProblemApplication.class, args);
-    }
-}
