--- gzhuoj-microservices/gzhuoj-user/src/main/java/com/gzhuoj/usr/UserApplication.java
@@ -1,13 +1,0 @@
-package com.gzhuoj.usr;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@MapperScan("com.gzhuoj.usr.mapper")
-public class UserApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(UserApplication.class, args);
-    }
-}
