+++ gzhuoj-services/gzhuoj-problem/src/main/java/com/gzhuoj/problem/ProblemApplication.java
@@ -1,0 +1,13 @@
+package com.gzhuoj.problem;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class ProblemApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ProblemApplication.class, args);
+    }
+}
