+++ gzhuoj-services/gzhuoj-user/src/main/java/com/gzhuoj/usr/UserApplication.java
@@ -1,0 +1,15 @@
+package com.gzhuoj.usr;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@MapperScan("com.gzhuoj.usr.mapper")
+@EnableDiscoveryClient
+public class UserApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(UserApplication.class, args);
+    }
+}
